Removed most To-Do List template HTML, CSS & JS
Deleted nearly all the JS having to do with the To-Do List project.
Confirmed some JS patterns for rendering objects to the UI and left
some JS code in the UI module as a template to use later on.

Imported some CSS style elements from other projects to potentially use
for this one and deleted the remaining CSS linked to the To-Do List
template.

Deleted most of the HTML elements from the To-Do List template. Kept
the HTML related to the search bar functionality and renamed some
selectors.