Add a new function for fetching location forecast
In the UI module, we removed comments in the sample render function
that were related to the old renderProject function in the To-Do List
template. They are no longer needed. We changed all naming conventions
of the previous form title variable to match the value of the new
search bar instead.

Removed the color black from all child elements under the content div
in the CSS since we're trying to introduce a light and dark mode. It
shouldn't be needed even if we don't end up having these modes.

In the app logic module, we wrote the function responsible for grabbing
the forecast of a given location from the Weather Crossing API. The
function simply logs the info to the console for now, but we'll need
to figure out how to call the function using a location without the use
of a form entry. In short, where should the name of the location go
when calling the function in order to get only the console to display
the info.

Imported the new getForecast function to the index module for use in
the search form logic later on.