Add more CSS styling and modify index module code
For the CSS, we added styling to all the buttons, commented in a
font-size option and a note about a max-width change in the search bar
styling and added a third setup that should center all content on page
load to the middle of the page. We also added a TODO regarding the
third setup to check and see how CSS Grid organizes those elements
initially before switching to that setup.

In the index module, we changed the searchFilter function to run the
getForecast function from our app logic module, then called
searchFilter in our search button logic. We are calling getForecast at
the top of the module to test the function first and make sure the
forecast info is generating in the console. We also commented in all of
the code from the cat GIF assignment to use as a potential template for
our API functions later on.