Add error messages as comments and more styling
Changed the favicon link in the HTML boilerplate in an attempt to get a
color version of the current favicon we were using to no avail. As
suspected, the grid styling does not look right in the CSS so we
changed over to the simple Flexbox centering setup for the time being.

Commented a note regarding the TypeError we are getting for the
headerTask which we knew would happen due to not doing anything with
the icon image from the To-Do List template yet. Commented the "access
to fetch" error message as well as the TypeError as a reminder of why
the getForecast function failed to fetch the forecast for the
location we entered when calling the function at the top of the index
module.