Reorganize CSS styling and template JS code
Reorganize icon image code from template To-Do List to the bottom of
the module and commented it out to possibly use it as a template later
on for the weather icons on the "results" page. Added a comment to the
getForecast function regarding a possible next step once the CORS issue
is figured out.

Reorganized styling elements for the content div in the CSS. Added in
further centering of the search bar and button elements within their
respective wrapper.