Modify getForecast function in debug attempts
Made several attempts to debug the getForecast function in the app
logic module to no avail. Tried adding a parameter to the function and
changing the API address by deleting the optional dates and changing
the [location] text to a template literal matching the function
parameter. Tried wrapping everything in the function into a try..catch
block. Tried setting the request mode to 'no-cors' as a previous error
message suggested. Get the same SyntaxErrors and the console doesn't
log the weather of "London, UK" as the sample location.