Resolve API fetch errors and refactor other code
In the app logic module, we resolved the errors in the getForecast
function whether they involved CORS, SyntaxErrors, TypeErrors or
otherwise. The function now properly grabs the weather info based on
the parameter given and shows it in the console. The 'try...catch'
block works properly as well.

Refactored the render function in the UI module from the old To-Do List
template to a function responsible for rendering the current
temperature to the UI based on the location the user requested in the
search bar of the app. Have yet to test this function in its current
form though. Made a note that even this new function is effectively a
template.

Refactored the search button logic in the index module to correspond
with the new render function in the UI module. Have yet to test this as
well to make sure everything renders correctly. Will do so in a future
update.

Made several comments throughout all modules for future TODO items,
templates for future code and general reminders of certain things to
work on.