Add TODO comments related to 7 day forecast func
Added TODO comments across all modules, with most of them being related
to how the 7 day forecast functions are going to work through each
module. Wrote a TODO comment in both the UI and app logic modules
related to fetching the correct GIF background based on weather
conditions. One for the actual fetch method and the other TODO for
rendering it to the page.

Wrote a comment in the CSS on a possible idea for how the transition of
the form elements are going to work once the user clicks the Search
button after typing in the location they want the weather for.