Modify forecastResults function in app logic
Added in several pieces of code as comments trying to get the correct
approach toward what the forecastResults function needs to do via the
pseudocode comments written just above it. The function is still not
fully fleshed out and will need to be refactored. Potentially need a
different approach than the 'for' loop currently there as right now
I'm not sure how to get the object the for loop returns on each cycle
into a new array that only contains the current temperature of a given
location along with the 7 day forecast.