Resolve error with variables calling getForecast
Reverted back to a previous version of the forecastResults function in
the app logic module in order to trace my steps back to how we got the
errors we were getting in the first place (see comments throughout the
code as the error notes are still there as of this commit). Figured out
that the issue was with todaysTemp and futureTemps variables having a
.response in the expression when it wasn't needed, causing the Type
Errors.

TypeErrors have been resolved but now the for loop doesn't work as it
should, which I expected. This is because returning the object as it
currently is ends the loop before it ever gets to the rest of the
other forecasts in the days array. So we likely need almost a totally
different approach involving nested arrays and/or objects. If the for
loop does work, it doesn't work in it's current implementation.