Rename and refactor the forecastResults function
Renamed forecastResults function to processData as that naming
convention fits what the function does slightly better based on TOP
instructions. Rewrote the name change in all necessary imports and
exports across the index and app logic modules.

Commented in another conditional statement under the already commented
for...of loop example inside of the processData function as a possible
solution to "skipping" the first object in the days array of the data
we're getting from the API. Deleted the initial for loop we came up
with due to the return statement inside of it ending the loop and no
way of getting the loop to continue in the function's current state.

Wrote a new object representing the "app state" of the weather info we
need specifically for this project (address/location, current temp,
forecast for the next 7 days with temp highs/lows each day).

Was able to get the new object futureForecast representing the "app
state" to return most of the specific info we need. The issue is the
entire 'days' array is still being returned instead of just objects 1
thru 7. We think there is either an issue with the conditionals inside
the loop or it's the 'for...of' loop itself and we need a different
loop (traditional 'for' loop?).