Resolve all errors for processData function
Was able to confirm that the errors in the processData that kept
returning the entire 15 day forecast in our newly returned object
instead of just the next 7 days was due to the conditional "if"
statements in our for...of loop and not our choice of loop itself.
What was happening was that the values that the for...of loop was
reading off of the days array were the objects themselves, not the
number representing the property, which is what the for...of loop does
(it doesn't read the property like a 'for...in' loop).

After several attempts at changing the conditionals, we were able to
land on '(day === futureTemps[#])'. with the # representing "0" for
skipping the first item in the array via 'continue' and # representing
"8" to end the loop with 'break' since at the point we have the 7 days
we need for the new object.

Commented in a possible refactor of the new object which would replace
the current variables/object combo we have, resulting in fewer lines of
code. Will test in the future.

Reorganized the pseudocode comments and placed each algorithm step in
its appropriate line of code within the processData function. The
pseudocode IS the description for each line, basically. Also commented
in a description for the getForecast function. Deleted all comments
that were notes used when trying to debug the errors in processData as
they are no longer needed.