Refactor new object returned in processData
Refactored some of the code in the 'for...of' loop that directly
created and returned the new object with only the data we need for the
app. Previously, variables were declared for each piece of data first
and then they were added as properties to the new object with no values.
Now, the object itself has the variable names as properties/keys with
each piece of data (previously as expressions) showing as the value for
each key. This reduces the lines of code we have.

Wrapped the entire processData in a try/catch block as well in order to
catch any possible future errors involving the API.