Reactivate searchFilter and search button logic
Reinstated the searchFilter function along with the search button
functionality in the index module. Revised the conditional in the
searchFilter function that deals with scenarios where the processData
function call fails (API or otherwise) to render an error message on
the web page.

Wrote some pseudocode for the search button logic and added each step
as comments at each relevant line. Changed the function call in the
button logic to the searchFilter function, but this may be a mistake
(see next paragraph).

Attempted to call the app logic functions in the searchFilter function
itself but has resulted in errors where one location is always showing
'undefined' no matter what. The app currently either can't get the
exact location at all (due to showing 'undefined') so it gets the next
closest location OR it DOES get the right location, but then the
getForecast function runs twice and grabs an 'undefined' location on
the second run. Have added several TODO comments throughout the app
logic module in regards to this.

It turns out the new object that processData would generate in the app
logic module needs to have the location of the forecast after all for
future rendering purposes. Declared the address being searched for as
a variable and added it as a property in the new object.

Removed comments in the UI module related to the 7 day forecast as they
were no longer needed. We have our 7 day forecast function in the app
logic module (processData).