Resolve location error showing as undefined
In the app logic module, I was able to debug the error where the
getForecast function's parameter would always show undefined when
called by the processData function. It turns out the solution was to
declare a global variable with no expression in the global scope so
both functions could access it as the target location from the user's
input when needed.

In the index module, I wrapped the searchFilter function into a
Promise. At first, this was in an attempt to debug the location error,
but it turned out to not work. I will leave the Promise in though
because the function is accessing an async function anyway, so the
Promise is very likely needed regardless.

Removed all TODO comments as well as other comments such as old code
that are no longer needed throughout both modules to clean the code up
a bit.