Implement pseudocode for getWeatherGIF function
Wrote the pseudocode step by step for the getWeatherGIF function, which
will be responsible for fetching a matching GIF from the GIPHY API for
the weather of the location that a user searches for. Ultimately the
function will render it to the page as well, which is why all of the
pseudocode will be moved from the app logic module to the UI module in
the near future.

Commented in some changes that will likely need to be made involving
how the getWeatherGIF function is going to access the 'outlook' object
returned by the processData function in the app logic module. The main
concern is accessing the icon text in order to get the right GIF.
Comments include a possible empty array to put the 'outlook' objects
into & returning the object a la closures/factory functions. There are
other comments with next steps to try first before exploring those.

Reactivated the renderCurrentTemp calls in the index module, both in
the search button logic and importing the function itself from the UI
module.

Added other TODO comments throughout the modules.