Add the icon set parameter for the current day
Declared the icon set parameter under the currentConditions object for
any fetched forecast as a variable for the future getWeatherGIF
function to use in order for the GIPHY API to fetch the correct GIF to
use as a background image on the page once rendered.

Moved all pseudocode for the future getWeatherGIF function over to the
UI module as the function will now be responsible for both fetching the
GIF from the GIPHY API and rendering the GIF to the page (these will
no longer be two separate functions).