Implement first draft for weather GIF fetch
Wrote up the first draft attempt at the function that will fetch the
background GIF related to the current weather condition that users
search for, called getWeatherGIF for the time being. Moved the variable
forecastDisplay to the global scope after realizing that the new
function will possibly need access to that DOM element to attach the
GIF to. Imported the processData function from the app logic module for
the same purpose as well. Will test the new function in a future update.