Implement first test of 'getWeatherGIF' function
Exported 'getWeatherGIF' from the UI module & imported it to the index
module for use in the search button event listener.

Reinstated 'processData' function call at the top of the index module
to test what it does on page load. Confirmed that it doesn't recognize
the previous string parameter in the 'getForecast' function call above
it. Leaving this the way it is for now, but I suspect there may be an
issue later on trying to get other functions to 'remember' what the
'getForecast' or 'processData' functions had for their location.

Wrote a TODO comment by the 'getWeatherGIF' function call regarding an
error where the function currently never gets called by the search
button event listener. Instead the listener stops after throwing the
current TypeError under the 'renderCurrentTemp' function in the UI
module, then finishing the 'searchFilter' function call.

Commented in the TypeError under the 'renderCurrentTemp' function in
question under the UI module.