Add and remove comments related to debug attempts
Added a TODO comment in the app logic module regarding the timing of
which line the TypeError in the weather GIF function (in the UI module)
is thrown. The pseudocode for the processData has also been reorganized
slightly.

Made 3 different attempts in the UI module at resolving the TypeError
related to the temperature text being rendered in the renderCurrentTemp
function, including changing it to async/await. Commented all the
attempts for reference. The summary (icon set parameter) variable in
the weather GIF function now throws the same TypeError and has been
recorded as a comment. Also added a TODO comment in the same function
to reorganize the conditional currently near the bottom of the function.

Removed the comment under the ELSE condition in the searchFilter
function under the index module as it has been tested for and is no
longer needed.