Resolve TypeError and broken GIF link bugs
Made several attempts to fix the TypeError bugs in the getWeatherGIF
function that were leading to broken GIF links. Some of these attempts
included:

Refactoring the searchFilter function to an async/await function in the
index module after making attempts to refactor it to a simple Promise
function. Did the same thing for the search button event listener after
making attempts to change it to a Promise function chaining more
Promises via '.then' as well as calling all the UI functions under one
'.then' Promise.

Making the same attempts at refactoring the getWeatherGIF function in
the UI module as the searchFilter function in the index module. Then
realized that bracket notation for the img element variable declarations
was needed in the switch statement inside the summonClip function
expression. Dot notation alone wasn't enough to get the GIF to render
based on weather conditions. Set a "limit" parameter on the fetch
request in that same function to stop the API request from sending back
50 different GIFs from GIPHY. This is a temporary fix for now. Wrapped
the whole getWeatherGIF in a try/catch block. This was an attempt to
resolve the errors mentioned before but did not resolve them. Leaving
the block in anyway to catch any other stray errors the app might bump
into.

Added a note about the change made to the return statement of the
processData function to make it a "factory function" in the app logic
module.

Added and removed general comments regarding timeline of what was worked
on in the app, what still needs to be worked on and debug attempts for
tracking purposes.