Pseudocode and refactor renderCurrentTemp function
Wrote some code in the search button event listener in the index module
to test out the behavior of the weather app when the Search button is
clicked after user input in the search bar while simultaneously
attempting to debug the renderCurrentTemp function in the UI module.
Specifically why nothing renders and why none of the code actually runs
(the browser/JS skips it almost entirely). Commented out the code
afterwards, then logged the attempts that were made on the function
call to renderCurrentTemp. Made some TODO comments for next steps on
the event listener.

Transitioned some previously written pseudocode from outside the text
editor into the renderCurrentTemp function in the UI module. Refactored
the function based on the pseudocode written. Logged any previous
refactoring attempts that were made to the function as comments as well
as adding some TODO comments for next steps. Will be testing the new
version of the function in a future update.