Resolve renderCurrentTemp errors & add UI elements
Was able to finally resolve all known errors that were popping up in
the renderCurrentTemp function in the UI module. The temperature, the
location icon and the related GIF all display to the UI. However, the
location itself was not displaying due to a TypeError that DevTools did
not catch that still had the location text equal to an expression that
was no longer in use. Resolved this error by updating the expression
itself.

Added more text elements to the renderCurrentTemp function that display
the weather condition summary to the UI. Created a wrapper to place all
the location elements under and appended all those related elements to
that wrapper. Refactored the append section of the function to reflect
these changes. Rewrote any of the TODO comment items that were completed
as pseudocode.