Modify comments and clean up some code
Changed the TODO comments in the getWeatherGIF function related to the
background GIF feature into BRANCH comments instead to reflect that
they will be worked on after the MVP release of the app.

Removed extra indentation in the renderCurrentTemp function that was
related to a previous conditional 'if' statement in the function that
has since been removed. Concatenated the Fahrenheit info to the
temperature expression line as the temperature was not displaying in
Fahrenheit or Celsius. It only displays the info given to us by the API
which is just a Fahrenheit number (likely based on location of user's
device where the input is searched from).