Create container for Fahrenheit & Celsius buttons
Created a container for the Fahrenheit and Celsius buttons similar to
the container for the search bar and Search button. Gave the container
some styling in order to space out those buttons from the search
content above and center them as well. Modified flex format of the
entire search container as well as the spacing on the search bar itself
in order to center it with the Search button.