Add and implement GPS icon image for location UI
Downloaded a new GPS icon image to use for our location image graphic
in order to replace the old Update icon from the To-Do List project
that was being used as a placeholder. Implemented the new icon image
both in the UI module itself and the renderCurrentTemp function that
is responsible for rendering it. Will test this out in the very near
future.

Shortened the placeholder text in the search bar so it all fits within
the search bar space in the HTML file.