Modify spacing between all rendered elements
Modified the spacing in the CSS for all rendered elements in the app.
Reorganized how the location and GPS icon render to give the app a
slightly cleaner look. Renamed the wrapper containing the Fahrenheit
and Celsius buttons in the HTML to "temperature-scales" as that is the
official definition of Fahrenheit and Celsius.