Modify weather summary text & button styling
Was able to resolve the vanilla text for the weather condition being
displayed on render when a location is searched for. The summary
text from the renderCurrentTemp function in the UI module now shows
the grammatically correct weather condition (all uppercase words,
no hyphens between the words). Resolved this with a combination of
the 'replace()' method in the renderCurrentTemp function and the
'text-transform' CSS property in our CSS file.

I've hidden the Fahrenheit and Celsius buttons via 'display: none'
in the CSS for now until we can figure out the toggle button logic
in the JS.

Added some comments in the getWeatherGIF function in the UI module
describing the issue with the GIFs being fetched and rendered for
any "clear" weather condition (it fetches a non-weather related
GIF every time). Currently don't have a GOOD fix for the issue so
leaving it as an issue to fix in a future branch for now.