Comment in example code for temp scale toggle
Added in some example code in the app logic module for some functions
responsible for toggling the temperature shown in the UI between the
Fahrenheit and Celsius scales. There are two different approaches that
have been commented in. Will pseudocode and develop the feature in a
future branch. For now, the Weather App is at MVP status and is
"complete".