Refactor summonClip function to fetch a random GIF
In the UI module under the getWeatherGIF function, we were able to
figure out how to get the GIPHY API to send a random GIF out of the 50+
GIF results that would come back from the API call. So we refactored the
summonClip function expression in the getWeatherGIF function to return
that randomGIF and render it to the UI instead of just limiting the GIF
fetch to the top result. Limiting the GIF to the top result was honestly
just a band-aid fix to a previous issue we had that caused the GIF to
not render at all whenever there was more than one GIF result returned.

In the app logic module, we commented in some example code for the
toggle Fahrenheit and Celsius functionality for the page to eventually
be used by the Fahrenheit and Celsius buttons.

Revised the note under the '.temperature-scales' selector in the CSS
file to swap to 'display: none;' if we can't figure out the JS code for
the Fahrenheit/Celsius toggling for some reason.