Resolve temperature scale toggling bug
Where we last left off:

Every time 'processData' was called (even via async/await when declaring
a variable using 'processData' as the expression), a new object would be
created due to 'processData' calling 'getForecast', which makes a new
API call to Weather Crossing. This caused the toggle function to only
work once (from Fahrenheit to Celsius) but never again. We couldn't
toggle the temperature back and forth in the console. We were trying to
figure out how to take existing objects/json we've already pulled from
an API and use/declare info from them without calling the API again. And
also how to take a new object we created from the API call/json, change
a value on one of its properties, then return that value/info without
fetching a new object via another API call (by calling 'processData' or
'getForecast').

Where we are now:

The solution we came up with was very similar to how we got processData
to remember the location info from getForecast. We declared 'data' as a
global variable (with no expression) to represent where the new object
from processData would "go to" after it was created (data = outlook).
Then we took the 'data' variable and used it in both toggle function
and I can confirm they do remember the new 'outlook' object that was
created from processData without having to make another API call. We
can now toggle between Fahrenheit and Celsius back and forth in the
console!

Will need to clean up the code and add the toggling to the UI itself
and then we will be done.